+++ lib/goods.rb
